+++ feed/sphinxcontrib/feed/__init__.py
@@ -21,6 +21,7 @@ def setup(app):
     app.add_config_value('feed_filename', 'rss.xml', 'html')
     
     app.connect('html-page-context', create_feed_item)
+    app.connect('html-page-context', inject_feed_url)
     app.connect('build-finished', emit_feed)
     app.connect('builder-inited', create_feed_container)
     app.connect('env-purge-doc', remove_dead_feed_item)
@@ -34,8 +35,12 @@ def create_feed_container(app):
     global feed_entries
     rss_fragment_path = os.path.realpath(os.path.join(app.outdir, '..', 'rss_entry_fragments'))
     feed_entries = FSDict(work_dir=rss_fragment_path)
-    # app.builder.env.feed_url = app.config.feed_base_url + \
-    #    app.config.feed_filename
+    app.builder.env.feed_url = app.config.feed_base_url + '/' + \
+        app.config.feed_filename
+    
+def inject_feed_url(app, pagename, templatename, ctx, doctree):
+    #We like to provide our templates with a way to link to the rss output file
+    ctx['rss_link'] = app.builder.env.feed_url #app.config.feed_base_url + '/' + app.config.feed_filename
     
 def create_feed_item(app, pagename, templatename, ctx, doctree):
     """
@@ -44,6 +49,7 @@ def create_feed_item(app, pagename, templatename, ctx, doctree):
     """
     global feed_entries
     import dateutil.parser
+    from absolutify_urls import absolutify
     date_parser = dateutil.parser.parser()
     metadata = app.builder.env.metadata.get(pagename, {})
     
@@ -65,14 +71,12 @@ def create_feed_item(app, pagename, templatename, ctx, doctree):
       'title': ctx.get('title'),
       'link': link,
       'unique_id': link,
-      'description': ctx.get('body'),
+      'description': absolutify(ctx.get('body'), link),
       'pubdate': pub_date
     }
     if 'author' in metadata:
         item['author'] = metadata['author']
-    feed_entries[nice_name(pagename, pub_date)] = item
-    #Additionally, we might like to provide our templates with a way to link to the rss output file
-    ctx['rss_link'] = app.config.feed_base_url + '/' + app.config.feed_filename
+    feed_entries[nice_name(pagename, pub_date)] = item    
 
 def remove_dead_feed_item(app, env, docname):
     """
@@ -84,7 +88,6 @@ def remove_dead_feed_item(app, env, docname):
     for name in feed_entries:
         if name.endswith(munged_name):
             del(feed_entries[name])
-        
 
 def emit_feed(app, exc):
     global feed_entries
@@ -93,6 +96,7 @@ def emit_feed(app, exc):
     feed_dict = {
       'title': app.config.project,
       'link': app.config.feed_base_url,
+      'feed_url': app.config.feed_base_url,
       'description': app.config.feed_description
     }
     if app.config.language: