+++ issuetracker/test_issuetracker.py
@@ -68,6 +68,58 @@ def pytest_funcarg__doc(request):
     return doc
 
 
+def test_get_github_issue_information(env):
+    info = issuetracker.get_github_issue_information(
+        'pyudev', 'lunaryorn', '2', env)
+    assert info == {'closed': True,
+                    'uri': 'https://github.com/lunaryorn/pyudev/issues/2'}
+
+
+def test_get_bitbucket_issue_information_resolved(env):
+    info = issuetracker.get_bitbucket_issue_information(
+        'synaptiks', 'lunar', '22', env)
+    assert info == {'closed': True,
+                    'uri': 'http://bitbucket.org/lunar/synaptiks/issue/22/'}
+
+
+def test_get_bitbucket_issue_information_invalid(env):
+    info = issuetracker.get_bitbucket_issue_information(
+        'synaptiks', 'lunar', '36', env)
+    assert info == {'closed': True,
+                    'uri': 'http://bitbucket.org/lunar/synaptiks/issue/36/'}
+
+
+def test_get_bitbucket_issue_information_duplicate(env):
+    info = issuetracker.get_bitbucket_issue_information(
+        'synaptiks', 'lunar', '42', env)
+    assert info == {'closed': True,
+                    'uri': 'http://bitbucket.org/lunar/synaptiks/issue/42/'}
+
+
+def test_get_google_code_issue_information_fixed(env):
+    info = issuetracker.get_google_code_issue_information(
+        'pytox', None, '2', env)
+    assert info == {
+        'closed': True,
+        'uri': 'http://code.google.com/p/pytox/issues/detail?id=2'}
+
+
+def test_get_google_code_issue_information_invalid(env):
+    info = issuetracker.get_google_code_issue_information(
+        'pytox', None, '5', env)
+    assert info == {
+        'closed': True,
+        'uri': 'http://code.google.com/p/pytox/issues/detail?id=5'}
+
+
+def test_get_google_code_issue_information_wontfix(env):
+    info = issuetracker.get_google_code_issue_information(
+        'pytox', None, '6', env)
+    assert info == {
+        'closed': True,
+        'uri': 'http://code.google.com/p/pytox/issues/detail?id=6'}
+
+
 def test_make_isssue_reference_resolver_invalid_reftype(
     app, env, resolver, node):
     node['reftype'] = 'spam'
@@ -113,11 +165,11 @@ def test_get_issue_information_called(
     app, env, resolver, node, get_issue_information):
     resolver(app, env, node, node[0])
     get_issue_information.assert_called_with(
-        'issuetracker', 'foobar', '10', env)
+        'issuetracker', 'foobar', '10', app)
     app.config.issuetracker_project = 'spam with eggs'
     resolver(app, env, node, node[0])
     get_issue_information.assert_called_with(
-        'spam with eggs', 'foobar', '10', env)
+        'spam with eggs', 'foobar', '10', app)
 
 
 def test_builtin_issue_trackers():
@@ -151,6 +203,16 @@ def test_issues_references(doc):
     _assert_text(em[0], 'see ')
     _assert_xref(em[1], '2')
     _assert_text(em[2], ' if you are #abc brave')
+
+
+def test_issues_references_too_many_groups(doc, config):
+    config.issuetracker_issue_pattern = r'(#)(\d+)'
+    transformer = issuetracker.IssuesReferences(doc)
+    with py.test.raises(ValueError) as exc_info:
+        transformer.apply()
+    error = exc_info.value
+    assert str(error) == ('issuetracker_issue_pattern must have '
+                          'exactly one group: %r' % ((u'#', u'1'),))
 
 
 def test_auto_connect_builtin_issue_resolvers_known_tracker(app):